+++ src/main/java/TubeSlice/tubeSlice/domain/oauth/OauthController.java
@@ -1,0 +1,11 @@
+package TubeSlice.tubeSlice.domain.oauth;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/v1/oauth")
+public class OauthController {
+}
