+++ src/main/java/TubeSlice/tubeSlice/domain/subtitle/Subtitle.java
@@ -1,5 +1,6 @@
-package TubeSlice.tubeSlice.domain;
+package TubeSlice.tubeSlice.domain.subtitle;
 
+import TubeSlice.tubeSlice.domain.video.Video;
 import jakarta.persistence.*;
 
 import java.time.LocalDateTime;
