+++ src/main/java/TubeSlice/tubeSlice/domain/user/SocialType.java
@@ -1,4 +1,4 @@
-package TubeSlice.tubeSlice.domain;
+package TubeSlice.tubeSlice.domain.user;
 
 public enum SocialType {
     KAKAO, NAVER
