+++ src/main/java/TubeSlice/tubeSlice/domain/user/Status.java
@@ -1,4 +1,4 @@
-package TubeSlice.tubeSlice.domain;
+package TubeSlice.tubeSlice.domain.user;
 
 public enum Status {
     ACTIVATE, INACTIVATE
