+++ src/main/java/TubeSlice/tubeSlice/domain/user/UserController.java
@@ -1,0 +1,11 @@
+package TubeSlice.tubeSlice.domain.user;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/v1/user")
+public class UserController {
+}
